+++ include/state_spaces/real_vector_space/RealVectorSpace.h
@@ -24,10 +24,10 @@ namespace base
 		virtual bool isValid(const std::shared_ptr<base::State> q) override;
 		// this one if cont. collision vs. discrete collision check is used
 		virtual bool isValid(const std::shared_ptr<base::State> q1, const std::shared_ptr<base::State> q2) override;
-		virtual double getDistance(const std::shared_ptr<base::State>q) override;
-		virtual std::tuple<double, std::shared_ptr<std::vector<Eigen::MatrixXd>>> getDistanceAndPlanes(const std::shared_ptr<base::State> q) override;
+		virtual float getDistance(const std::shared_ptr<base::State>q) override;
+		virtual std::tuple<float, std::shared_ptr<std::vector<Eigen::MatrixXf>>> getDistanceAndPlanes(const std::shared_ptr<base::State> q) override;
 
-		std::shared_ptr<base::State> interpolate(const std::shared_ptr<base::State> q1, const std::shared_ptr<base::State> q2, double step, double D) override;
+		std::shared_ptr<base::State> interpolate(const std::shared_ptr<base::State> q1, const std::shared_ptr<base::State> q2, float step, float D) override;
 		bool equal(const std::shared_ptr<base::State> q1, const std::shared_ptr<base::State> q2) override;
 		virtual std::shared_ptr<base::State> randomState() override;
 	protected:
