+++ src/state_spaces/real_vector_space/RealVectorSpace.cpp
@@ -33,12 +33,12 @@ bool base::RealVectorSpace::isValid(const std::shared_ptr<base::State> q)
 	return true;
 }
 
-double base::RealVectorSpace::getDistance(const std::shared_ptr<base::State> q)
+float base::RealVectorSpace::getDistance(const std::shared_ptr<base::State> q)
 {
 	return 0;
 }
 
-std::tuple<double, std::shared_ptr<std::vector<Eigen::MatrixXd>>> base::RealVectorSpace::getDistanceAndPlanes(const std::shared_ptr<base::State> q)
+std::tuple<float, std::shared_ptr<std::vector<Eigen::MatrixXf>>> base::RealVectorSpace::getDistanceAndPlanes(const std::shared_ptr<base::State> q)
 {
 	return {0, nullptr};
 }
@@ -52,7 +52,7 @@ std::shared_ptr<base::State> base::RealVectorSpace::randomState()
 
 // D (optional parameter) is the distance between q1 and q2
 std::shared_ptr<base::State> base::RealVectorSpace::interpolate(const std::shared_ptr<base::State> q1, 
-																const std::shared_ptr<base::State> q2, double step, double D)
+																const std::shared_ptr<base::State> q2, float step, float D)
 {
 	std::shared_ptr<base::State> q_t = std::make_shared<base::RealVectorSpaceState>(dimensions);
 	Eigen::VectorXf eig;
@@ -79,8 +79,8 @@ std::shared_ptr<base::State> base::RealVectorSpace::interpolate(const std::share
 
 bool base::RealVectorSpace::equal(const std::shared_ptr<base::State> q1, const std::shared_ptr<base::State> q2)
 {
-	double d = (q1->getCoord() - q2->getCoord()).norm();
-	double stateEqualityThreshold = 1e-6; // TODO: needs to be obtained from configuration file
+	float d = (q1->getCoord() - q2->getCoord()).norm();
+	float stateEqualityThreshold = 1e-6; // TODO: needs to be obtained from configuration file
 	if (d < stateEqualityThreshold)
 		return true;
 	return false;
@@ -89,8 +89,8 @@ bool base::RealVectorSpace::equal(const std::shared_ptr<base::State> q1, const s
 bool base::RealVectorSpace::isValid(const std::shared_ptr<base::State> q1, const std::shared_ptr<base::State> q2)
 {
 	int numChecks = 10;
-	double D = (q2->getCoord() - q1->getCoord()).norm();
-	for (double t = 1./numChecks; t <= 1; t += 1./numChecks)
+	float D = (q2->getCoord() - q1->getCoord()).norm();
+	for (float t = 1./numChecks; t <= 1; t += 1./numChecks)
 	{
 		std::shared_ptr<base::State> q_t = interpolate(q1, q2, t, D);
 		if (q_t == nullptr)
